services/ai-backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
google-cloud-aiplatform>=1.38.0
pydantic>=2.9.0
python-multipart>=0.0.9
pypdf>=3.17.0
python-dotenv>=1.0.0
google-genai>=0.2.0
transformers>=4.35.0
torch>=2.1.0 --index-url https://download.pytorch.org/whl/cpu
numpy<2.0.0
google-cloud-firestore>=2.13.0
google-cloud-storage>=2.13.0
